@@ -25,87 +25,84 @@
     "postinstall": "electron-builder install-app-deps"
   },
   "dependencies": {
-    "@angular/animations": "^7.0.0-beta.0",
-    "@angular/cdk": "^6.4.2",
-    "@angular/common": "^7.0.0-beta.0",
-    "@angular/compiler": "^7.0.0-beta.0",
-    "@angular/core": "^7.0.0-beta.0",
-    "@angular/platform-browser": "^7.0.0-beta.0",
-    "@angular/platform-browser-dynamic": "^7.0.0-beta.0",
-    "@angular/router": "^7.0.0-beta.0",
-    "@fortawesome/angular-fontawesome": "^0.1.1",
-    "@fortawesome/fontawesome-svg-core": "^1.2.2",
-    "@fortawesome/free-brands-svg-icons": "^5.2.0",
-    "@fortawesome/free-regular-svg-icons": "^5.2.0",
-    "@fortawesome/free-solid-svg-icons": "^5.2.0",
-    "@ngrx/effects": "^6.1.0",
-    "@ngrx/router-store": "^6.1.0",
-    "@ngrx/store": "^6.1.0",
-    "@ngrx/store-devtools": "^6.1.0",
-    "@swimlane/ngx-charts": "^9.0.0",
+    "@angular/animations": "^8.1.1",
+    "@angular/cdk": "^8.0.2",
+    "@angular/common": "^8.1.1",
+    "@angular/compiler": "^8.1.1",
+    "@angular/core": "^8.1.1",
+    "@angular/platform-browser": "^8.1.1",
+    "@angular/platform-browser-dynamic": "^8.1.1",
+    "@angular/router": "^8.1.1",
+    "@fortawesome/angular-fontawesome": "^0.4.0",
+    "@fortawesome/fontawesome-svg-core": "^1.2.19",
+    "@fortawesome/free-brands-svg-icons": "^5.9.0",
+    "@fortawesome/free-regular-svg-icons": "^5.9.0",
+    "@fortawesome/free-solid-svg-icons": "^5.9.0",
+    "@ngrx/effects": "^8.1.0",
+    "@ngrx/router-store": "^8.1.0",
+    "@ngrx/store": "^8.1.0",
+    "@ngrx/store-devtools": "^8.1.0",
+    "@swimlane/ngx-charts": "^12.0.1",
     "common-tags": "^1.8.0",
-    "core-js": "^2.5.7",
-    "d3": "^5.5.0",
-    "date-fns": "^2.0.0-alpha.10",
-    "debug": "^3.1.0",
+    "core-js": "^3.1.4",
+    "d3": "^5.9.7",
+    "date-fns": "^2.0.0-alpha.16",
+    "debug": "^4.1.1",
     "dexie": "^2.0.4",
     "electron-contextmenu-middleware": "^1.0.3",
     "electron-exif-rotate": "^0.2.1",
     "electron-image-menu": "^2.0.0",
-    "electron-window-state": "^4.1.1",
-    "keytar": "^4.2.1",
-    "lodash": "^4.17.10",
+    "electron-window-state": "^5.0.3",
+    "keytar": "^4.11.0",
+    "lodash": "^4.17.14",
     "monzolib": "^0.1.0",
-    "p-iteration": "^1.1.7",
-    "reflect-metadata": "^0.1.12",
-    "request": "^2.87.0",
-    "request-promise-native": "^1.0.5",
-    "rxjs": "^6.2.2",
-    "rxjs-compat": "^6.2.2",
-    "zone.js": "^0.8.26"
+    "p-iteration": "^1.1.8",
+    "reflect-metadata": "^0.1.13",
+    "rxjs": "^6.5.2",
+    "rxjs-compat": "^6.5.2",
+    "zone.js": "^0.9.1"
   },
   "devDependencies": {
-    "@angular/cli": "^6.2.0-beta.1",
-    "@angular/compiler-cli": "^7.0.0-beta.0",
-    "@ngtools/webpack": "^6.2.0-beta.1",
-    "@types/common-tags": "^1.4.0",
-    "@types/d3": "^5.0.0",
-    "@types/debug": "0.0.30",
-    "@types/electron-config": "^0.2.1",
+    "@angular/cli": "^8.1.1",
+    "@angular/compiler-cli": "^8.1.1",
+    "@ngtools/webpack": "^8.1.1",
+    "@types/common-tags": "^1.8.0",
+    "@types/d3": "^5.7.2",
+    "@types/debug": "^4.1.4",
+    "@types/electron-config": "^3.2.2",
     "@types/electron-devtools-installer": "^2.2.0",
-    "@types/electron-window-state": "^2.0.31",
-    "@types/lodash": "^4.14.116",
-    "@types/node": "^10.5.6",
-    "@types/request-promise-native": "^1.0.15",
-    "@types/sharp": "^0.17.9",
-    "copy-webpack-plugin": "^4.5.2",
-    "copyfiles": "^2.0.0",
+    "@types/electron-window-state": "^2.0.33",
+    "@types/lodash": "^4.14.136",
+    "@types/node": "^12.6.2",
+    "@types/sharp": "^0.22.2",
+    "copy-webpack-plugin": "^5.0.3",
+    "copyfiles": "^2.1.1",
     "cross-env": "^5.2.0",
-    "css-loader": "^1.0.0",
+    "css-loader": "^3.0.0",
     "devtron": "^1.4.0",
-    "eclint": "^2.8.0",
-    "electron": "^3.0.0-beta.4",
-    "electron-builder": "^20.26.1",
+    "eclint": "^2.8.1",
+    "electron": "^4.2.6",
+    "electron-builder": "^21.0.15",
     "electron-devtools-installer": "^2.2.4",
-    "electron-reload": "^1.2.5",
-    "file-loader": "^1.1.11",
+    "electron-reload": "^1.4.0",
+    "file-loader": "^4.0.0",
     "html-loader": "^0.5.5",
     "html-webpack-plugin": "^3.2.0",
-    "husky": "^0.14.3",
+    "husky": "^3.0.0",
     "json-types": "^1.0.0",
-    "lint-staged": "^7.2.0",
-    "mini-css-extract-plugin": "^0.4.1",
-    "postcss": "^7.0.2",
-    "postcss-color-gray": "^4.1.0",
-    "postcss-easing-gradients": "^3.0.0",
-    "postcss-loader": "^2.1.6",
-    "postcss-preset-env": "^5.3.0",
-    "prettier": "^1.14.0",
-    "style-loader": "^0.21.0",
+    "lint-staged": "^9.2.0",
+    "mini-css-extract-plugin": "^0.7.0",
+    "postcss": "^7.0.17",
+    "postcss-color-gray": "^5.0.0",
+    "postcss-easing-gradients": "^3.0.1",
+    "postcss-loader": "^3.0.0",
+    "postcss-preset-env": "^6.7.0",
+    "prettier": "^1.18.2",
+    "style-loader": "^0.23.1",
     "to-string-loader": "^1.1.5",
-    "tslib": "^1.9.3",
-    "typescript": "^2.9.2",
-    "webpack": "^4.16.4",
-    "webpack-cli": "^3.1.0"
+    "tslib": "^1.10.0",
+    "typescript": "^3.5.3",
+    "webpack": "^4.35.3",
+    "webpack-cli": "^3.3.5"
   }
 }
