@@ -1,13 +1,15 @@
-import { ChangeDetectionStrategy, Component } from '@angular/core'
+import { ChangeDetectionStrategy, Component, Input } from '@angular/core'
 
 @Component({
   selector: 'm-btn',
   template: `
-    <button>
+    <button [type]="type">
       <ng-content></ng-content>
     </button>
   `,
   styleUrls: ['btn.component.css'],
   changeDetection: ChangeDetectionStrategy.OnPush
 })
-export class BtnComponent {}
+export class BtnComponent {
+  @Input() type = 'submit'
+}
